* add arbitrum chain configuration and addresses

* move aave health factor from shapshot to AaveAccountSnapshot.cs table

* Add LiquidationLtv for MorphoMarketPositionSnapshot.cs

* add wallet and network to MorphoPositionsStatus

* add migrations and morpho state message

* add lltv to aave snapshots

* add migrations

* add aave health factor to /status command

* update lltv percent

+++ src/modules/aave/CryptoWatcher.Modules.Aave.Infrastructure/Client/UiPoolDataProvider/Contracts/ReservesData/BaseCurrencyInfo.cs
@@ -16,5 +16,5 @@ public class BaseCurrencyInfo
     public BigInteger NetworkBaseTokenPriceInUsd { get; set; }
 
     [Parameter("uint8", "networkBaseTokenPriceDecimals", 4)]
-    public byte NetworkBaseTokenPriceDecimals { get; set; }
+    public BigInteger NetworkBaseTokenPriceDecimals { get; set; }
 }
